Move reusable modules in main concerns directory
