+++ app/presenters/concerns/contact_details_api_data.rb
@@ -1,0 +1,28 @@
+module ContactDetailsAPIData
+  def contact_details
+    if application_form.international_address?
+      address_line1 = application_form.address_line1 || application_form.international_address
+
+      {
+        phone_number: application_form.phone_number,
+        address_line1: address_line1,
+        address_line2: application_form.address_line2,
+        address_line3: application_form.address_line3,
+        address_line4: application_form.address_line4,
+        country: application_form.country,
+        email: application_form.candidate.email_address,
+      }
+    else
+      {
+        phone_number: application_form.phone_number,
+        address_line1: application_form.address_line1,
+        address_line2: application_form.address_line2,
+        address_line3: application_form.address_line3,
+        address_line4: application_form.address_line4,
+        postcode: application_form.postcode,
+        country: application_form.country,
+        email: application_form.candidate.email_address,
+      }
+    end
+  end
+end
